CodeGen from PR 19983 in Azure/azure-rest-api-specs Merge 36f190aa6f5cd4485bb3ec1f9632a3e7dcc6b9e3 into 64d1545fcbb9c6b4f41e2fd1fd2f3e7cb421772d

+++ sdk/iotcentral/azure-resourcemanager-iotcentral/src/main/java/com/azure/resourcemanager/iotcentral/fluent/models/AppTemplateInner.java
@@ -40,7 +40,7 @@ public final class AppTemplateInner {
      * The order of the template in the templates list.
      */
     @JsonProperty(value = "order", access = JsonProperty.Access.WRITE_ONLY)
-    private Float order;
+    private Integer order;
 
     /*
      * The description of the template.
@@ -101,7 +101,7 @@ public final class AppTemplateInner {
      *
      * @return the order value.
      */
-    public Float order() {
+    public Integer order() {
         return this.order;
     }
